Add config option to use the system bus.
This commit adds a config option to use the system bus instead of the session bus. This change makes it possible to expose Mopidy through an MPRIS interface even when it's run in its own user without a session bus.

This commit relates closely to issue #9, and might solve it as far as possible within the scope of this extension.

The check for whether the `$DISPLAY` X11 environment variable is set is removed from the `validate_environment(self)` method in the main Extension class. Dbus will take care of this error, and there is already logic to handle it in the code that deals with dbus. the error will now be printed like this:

    WARNING  MPRIS frontend setup failed (org.freedesktop.DBus.Error.NotSupported: Unable to autolaunch a dbus-daemon without a $DISPLAY for X11)

This error message is equally verbose and easy to spot.

Signed-off-by: totokaka <mail@totokaka.io>